requirements.txt
matplotlib
Pillow
numpy
opencv-python